Improve usage of class loaders
Class loaders are not discarded after use, but kept cached, so that
subsequent use of the same class loader can reuse already loaded (and
possibly JIT-compiled) classes.

Tool class loaders are separated from combined tool-project class
loaders, so that tool class loaders can be reused when the same tool
is ran on different projects.

Combined tool-project class loader is not even created unless
particular tool is annotated with @ProjectClassScope, which means that
the tool needs a class loader that includes project classes.